requirements.txt
amqp==5.3.1
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
blessed==1.20.0
celery==5.4.0
cffi==1.17.1
channels==3.0.5
channels_redis==3.3.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.0
daphne==3.0.2
dj-database-url==2.3.0
Django==4.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-picklefield==3.2
django-q==1.3.9
django-redis==5.4.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jinxed==1.3.0
kombu==5.4.2
msgpack==1.1.0
packaging==24.2
pillow==10.4.0
prompt_toolkit==3.0.49
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.3.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==3.5.3
service-identity==24.2.0
setuptools==75.5.0
six==1.16.0
sqlparse==0.5.1
Twisted==24.11.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
zope.interface==7.2
django-csp==3.8
django-extensions==3.2.3
paho-mqtt==1.6.1